Change PageLayout interactions to use Playwright.
